system/Question_Answer_Engine/requirements.txt
fastapi
uvicorn
aiohttp
python-dotenv
langchain
langchain_community
sentence-transformers
faiss-cpu
huggingface-hub
pydantic
bs4
PyPDF2
